main오류수정_호성

+++ src/main/java/com/keduit/helloworld/repository/CouponRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.keduit.helloworld.entity.Coupon;
 
-public interface CouponRepository extends JpaRepository<Coupon, Integer>{
+public interface CouponRepository extends JpaRepository<Coupon, Long>{
 
 	
 }
