+++ src/main/java/com/keduit/helloworld/serviceImpl/CouponServiceImpl.java
@@ -21,7 +21,7 @@ public class CouponServiceImpl implements CouponService{
 
 	@Override
 	/** 쿠폰 번호 등록 */
-	public Integer register(CouponDTO dto) {
+	public Long register(CouponDTO dto) {
 		Coupon entity = DtoToEntity(dto);
 		
 		couponRepository.save(entity);
@@ -39,7 +39,7 @@ public class CouponServiceImpl implements CouponService{
 
 	@Override
 	/** 쿠폰 삭제 */
-	public void remove(Integer couponNum) {
+	public void remove(Long couponNum) {
 
 		couponRepository.deleteById(couponNum);
 	}
